requirements.txt
python
flask
jinja
flask-wtf
Flask-SQLAlchemy
python-dotenv
email-validator